version 2.35_p2

+++ lib/randomCo.hpp
@@ -74,7 +74,7 @@ namespace randomCo
   
     // Combine time with counter to vary each call
     auto now = std::chrono::high_resolution_clock::now().time_since_epoch().count();
-    uint32_t state = static_cast<uint32_t>(now) ^ (counter++);
+    uint64_t state = static_cast<uint64_t>(now) ^ (counter++);
   
     // Xorshift32
     state ^= state << 13;
